Make typestate pattern for aggregator and enable custom context strings.
This shifts all the errors in signature aggregation which are the fault of the
aggregator onto the aggregator themselves, as they need to ensure they have the
proper state before they can move on to the actual aggregation protocol, leaving
only the parts of the aggregation protocol which are the fault of the signers
(i.e. a malformed partial signature) at play.

Also the context string for the message hash can be application defined.